Create a common tool for ocarina Python to takle error
